Extract client proxy to separate class.
